Move module :library:process-testing -> :test-api
