+++ .github/workflows/CI.yml
@@ -105,7 +105,7 @@ jobs:
         with:
           emulator-boot-timeout: 300 # 5 minutes
           api-level: ${{ matrix.api-level }}
-          script: ./gradlew :library:process-testing:connectedCheck -PKMP_TARGETS="ANDROID,JVM"
+          script: ./gradlew connectedCheck -PKMP_TARGETS="ANDROID,JVM"
 
       - name: Upload Test Reports
         uses: actions/upload-artifact@v4
