+++ settings.gradle.kts
@@ -38,8 +38,9 @@ if (CHECK_PUBLICATION != null) {
 } else {
     listOf(
         "process",
-        "process-testing",
     ).forEach { module ->
         include(":library:$module")
     }
+
+    include("test-api")
 }
