+++ test-api/src/androidMain/AndroidManifest.xml
